Following
Create a relationship model/table which belongs to follower and following, which are both aliases for user. Implement a has-many relationship for active and passive on user, both referring to the Relationship class. Set user associations for following and followers. Create useful helper methods in the User model for following.